Added controls to vuex, as a state property allControls.
Currently they are randomly populated by a snippet in the App.vue mounted() hook. The controls therein are NOT the final representation; that's a work in progress being conducted elsewhere. These are intended SOLELY to stop blocking others work testing visual effects.

Statistics for these controls can be found/added as submodules to the vuex store. See "status_counts.js" for an example.

The old data in the vuex store, used for themeing the app, can now be found in the "store/theme" module. It also holds active user info but I'm pretty sure we're ditching that.

Look to results.vue for an example of how to use data from the vuex store. :statistic="$store.getters['statusCounts/notReviewed']"
This binds the CountCard property statistic to the getter value notReviewed from the statusCounts module, found in the vuex store.