ENH: Add support for callbacks on JSON.parse error
Add callback option to dispatcher for load_file, only one so far,
that is for if JSON.parse() errors out due to syntax error.

A better way to do this is to write a Toast component, and send
TOAST_MESSAGES through vuex whenever an error happens, but this works
for testing now.

Related to #97

Signed-off-by: Luke Malinowski <lmalinowski@mitre.org>